Add IFC support for old petSD and the new pet microSD
The code tries to read the die revision of the ethernet
interface chip ENC28J60. If it fails, it assumes that this chip
is not present and that IFC is connected to a port pin that would
have been used as ethernet interrupt otherwise.

Having neither ethernet chip nor IFC patch on the old petSD
does not harm. It will simply ignore IFC as it always did.

IFC is connected to port PD3 on the pet microSD by default.
Have a look at http://petsd.net for instructions how to patch
in two wires to make IFC usable for the old petSD as well.