ieee488: fix an ATN race condition
If ATN was released and set again in a very short time, it could
happen that the PET got stuck waiting forever.

NRFD set high (to indicate ready for data) got overwritten by the
ATN acknowledge interrupt to NRFD low (not ready for data).
By consequence, the PET got stuck because it waited without
timeout for NRFD low.

The current code prevents such race conditions by not checking
ATN directly but indirectly, in combination with the ATN interrupt
routine.