Bus sleep while menu active
As long as the menu system is active, the bus interface is detached.
The red LED lights constantly as a reminder for disk operations
taken out of service.

During this state, you are able to access drives with the same
device number which would conflict normally, e.g. to change
their device number to a non-conflicting value.