Bugfix: use 11 as default device address
Bugfix: if the device address couldn’t get read back from the
EEPROM memory, it was accidentally set to zero.

This bugfix uses 11 as a new default value. This value was chosen
because it doesn’t conflict with 8 to which internal drives are usually
set, such as for example the CBM 8296D.

A new config value CONFIG_DEFAULT_ADDR is introduced
(see configs/config-example) to compile for default addresses
other than 11.