Send debug messages to UART, LCD or both
Traditionally, debug output was sent to the serial line
if CONFIG_UART_DEBUG=y was given. This is not applicable
if we want to transfer data over the serial line with the SPSP
Serial PET/Server protocol so we need a mechanism to
output debug messages to the LCD, the UART or both, depending
on the current debugging needs.

Debug output can be sent to

     - the UART if CONFIG_UART_DEBUG=y is given
     - the LCD if CONFIG_DEBUG_MSGS_TO_LCD=y is given
     - or both if both options are enabled

All higher level debug output functions are based on two core functions
debug_putc() and debug_flush() so the former is used for distribution.