Fix IEC fastloader for petSD+
The AVR fastloader code is hardcoded for 8 MHz and running
the device at other speeds breaks the timing and thus the
support of fastloaders.

But the device can not run at 8 MHz because that would be too
slow to respond in time to the IEEE-488 ATN signal.

To fix this, a CPU clock prescaler is introduced to run at 8 MHz
when in IEC mode and at 16 MHz when in IEEE-488 mode.

Replace the crystal on your petSD+ with a 16 MHz type to benefit
from this commit. However, take greatest care not to damage the
PCB when you exchange the crystal!