Fix SPI speed during card init for petSD+
The divisors for slow and fast speed were mixed up so
that the card was initialized way too fast. For some cards.
Maybe. At least according to the specs from many years ago.
In practice, this bug doesn't seem to be much of a problem,
it was found just by random.